Fix crash with genva() in async generators
Summary:
genva() uses a new AwaitAll opcode, which wasn't handled properly inside
opcodeControlFlowInfo(). Since async generators are not yet supported by JIT,
the opcode was handled by the interpreter using InterpOne instead of
InterpOneCF, ignoring the new PC/SrcKey and continuing execution instead of
transfering control to the caller (eager execution) or the asio scheduler.

Let's prevent this by adding an assert to InterpOne, add the opcode to the
list and audit all CF opcodes, which revealed that FCallAwait was also missing.

Also, handle AwaitAll in region selection similar to Await and FCallAwait.

Reviewed By: ricklavoie, paulbiss

Differential Revision: D6230160

fbshipit-source-id: af3e82738952544e1ea81dcd11d5a825fd13774d