Provide default values if Futures in Informant fail
Summary:
When one of the Informant's futures throws, it gets caught by the monitor's main
run loop, which logs the uncaught exception, and hits the informant again, which
triggers the same exception to be throw again.

This results in an infinite throw-catch loop and exploding log file.

This fixes that

Differential Revision: D6229165

fbshipit-source-id: 8b4f0b8b0dfdb7627d757af2768bd277259a2cc0