Fix error in syntax folder
Summary: While attempting to implement some logic in the positioned syntax tree I noticed that this auto-generated method was wrong when the argument was a list.  In the current codebase we never call this method with a list, so the error is invisible. But I figured that we might as well fix it now.

Reviewed By: vladima

Differential Revision: D6214483

fbshipit-source-id: 283faed33affef463b90551cc5d83f65ceaa86fa