Tweak parser for cast / lambda / paren disambiguation
Summary:
This is a micro-optimization for a common cause of long backtracks in the parser. We need to be able to tell

    (T)$x
    (T $t)==>$t
    ($t, $u)==>$t
    ($t + $u)

and so on apart when we encounter a (. The technique we would use is to first, see if it meets the requirements of a cast; if yes, reparse it as cast. If not, see if it meets the requirements of a lambda signature. If yes, reparse it as a lambda signature. Otherwise, its a parenthesized expression.

This means that if it is a cast or a lambda, we end up parsing at least three tokens multiple times; in the case of a lambda signature we can be re-parsing a dozen tokens or more.

We now do not backtrack if it really is a cast or lambda.

Differential Revision: D5971544

fbshipit-source-id: 1cffc1dd2e9c7c78e23da0633b94023136071278