Catch unexpected exceptions while writing the repo
Summary:
One of the hhbbc crashes last week manifested as "std::terminate
called with no active exception", which gave us very little to go on.

The reason was that we were throwing while writing the repo, which
attempted to destroy both dispatcherThread and wp_thread while they
were still running (which was in turn because memory corruption was
causing a string to appear to exceed the serialization limit).

This adds better logging for that situation, to make it easier to
figure out whats going on (note that it will still std::terminate in
exactly the same way as before).

Reviewed By: ricklavoie

Differential Revision: D6242074

fbshipit-source-id: cdb8046cf292db6fffe2c7e55e7a43fbc354cb02