Properly set [v]array flag in header for casts
Summary:
Properly set/clear the varray flag in ArrayData for casts to/from varrays. As a
result of this, CastArray and CastVArray are no longer no-ops on array inputs,
so deal with that in the JIT and HHBBC as appropriate. Add more assertions to
verify the sanity of the [d,v]array flags.

Reviewed By: mxw

Differential Revision: D5756049

fbshipit-source-id: 20d2540c2c684ede3b0502e5696975544be38c55