Add d/varray awareness to HHBBC's type-system
Summary:
Add support for understanding the d/varray distinction in HHBBC's type-system. I
had originally wished to not do this, but without it, we have to perform too
many pessimizations. For instance, its no longer safe to generate a static array
when we know all the values because we don't know if it should be a static
varray, darray, or none. By adding support for this in the HHBBC type-system, we
can know exactly which array to create.

Add new types to HHBBC: T[C,S]PArr[N,E], T[C,S]VArr[N,E], and T[C,S]DArr[N,E],
where C/S/N/E have the usual meanings, along with the typical unions of
such. TPArr means a "plain" PHP array, neither darray nor varray. TVArr means a
varray. TDArr means a darray. TArr and family is now a union of TPArr, TVArr,
and TDArr, and represents situations where we don't statically know if an array
is a d/varray or not. The new types required expanding trep to 64-bits, but that
does not seem to have any real memory impact.

Now that we can statically know if an array is a d/varray or not, we can restore
the ability to optimize away CastArray, CastVArray, and CastDArray.

Add the type modifier "loosen_dvarrayness", which raises any specific TPArr,
TVArr, or TDarr type to be just TArr (while keeping any static or emptiness
bits). This is needed because in certain contexts (such as comparisons), the
d/varray-ness of an array does not matter. Also add "promote_varray", which
changes any TVArr bits to be TDArr, and represents operations where we
definitely know a varray will promote to a darray. Likewise add
"maybe_promote_varray", which adds a TDArr bit for every matching TVArr bit, and
represents operations where we might promote a varray to a darray (or might not,
so we need to keep the TVArr bits).

Reviewed By: markw65

Differential Revision: D6107522

fbshipit-source-id: 7db90e798363406de60dbfd2a1f53d8aaabe7adc