Add varray/darray checking IR ops
Summary:
Add IR ops (CheckVArray and CheckDArray) which can check if an array is a varray
or darray. Use these to implement is_varray() and is_darray() using IsTypeL and
IsTypeC just like other type-checks. I chose to implement these as Check* ops
which branch, rather than Is* ops which return a boolean. This allows us to
refine the type of the array if the check succeeds, since a varray is always
packed, and a darray is always mixed.

In addition, use them to short-circuit no-op varray and darray casts without
calling a runtime helper.

Reviewed By: paulbiss

Differential Revision: D6108032

fbshipit-source-id: 204f19514bea2202c92f7f991279d9ebc5ddab31