Add support for distinguishing d/varray type annotations
Summary:
We have a need to be able to distinguish d/varray type annotations because we
want to be able to warn when there's a mismatch between a type hint and a
value. However, darrays and varrays have KindOfArray just like normal
arrays. So, instead of using a new precise AnnotType, invent three new
AnnotMetaTypes (for varray, darray, and varray_or_darray) and form AnnotTypes
from them. Special case in the necessary places for these new AnnotMetaTypes to
behave like AnnotType::Array.

When used in builtin return types, these AnnotTypes imply the return type is
definitely that flavor of d/varray, so utilize that information in HHBBC and the
JIT.

However, for normal type-hints, HHBBC cannot utilize this information, because
the d/varray-ness of inputs or return values will not be hard enforced. So, have
HHBBC loosen away the dvarrayness of the information in the type-constraints
where applicable.

Since we don't have any builtins currently with d/varray type annotations. Add a
few dummy ones in the intrinsics namespace for testing.

Reviewed By: markw65

Differential Revision: D6147580

fbshipit-source-id: bec689208d0b8c71e3fb950bf38f34cbd2c87c07