Make variadics be varrays
Summary:
Change variadic parameters to be varrays. We were using PackedArray::MakePacked
in a few places to create the variadic array. This meant that
HackArrCompatNotices were not being generated if we attempted to bind the
variadic params by ref. Instead use the ArrayInit classes. After this change,
MakePacked, MakeVArray, and MakeVec all are only used on lists of cells, so
change the function signature to match (and enforce with asserts). Add some
handling of exceptions when we're setting up the variadic array where we
couldn't throw before.

Reviewed By: mxw

Differential Revision: D6163083

fbshipit-source-id: eaa1667ebf81474738bc130d3a57363b493b1d45