Add HackArrCompat notices for when a varray promotes to a darray
Summary:
Add a HackArrCompat (controlled by a new HackArrCompatPromoteNotices runtime
option) for when a varray promotes to a darray (because it can no longer remain
packed). This provides hints that a varray shouldn't be a varray because its
used at some point like a darray. In addition, provide a notice for when a
varray is implicitly appended to by doing a set of the "next" index. If a varray
is a vec, this won't work, so its a behavorial change.

Reviewed By: alexeyt

Differential Revision: D6170849

fbshipit-source-id: 3fd931006711b3023d6ee20c11e02f653de07fbd