Add HackArrCompat notices for when a d/varray type-hint is violated
Summary:
Add new HackArrCompat notices (controlled by a new HackArrCompatTypeHintNotices
runtime option) for when an array is passed (or returned) from a function, and
the type-hint and the array does not match with regards to d/varray-ness. The
same notices are raised when likewise a type constraint is violated on a builtin
(but not the return type). This is necessary because if varray becomes a vec,
and darray becomes a dict, they can no longer be passed to functions expecting
each other interchangeably.

Reviewed By: mxw

Differential Revision: D6171057

fbshipit-source-id: 8c878d97869b5df25ad76fb4cb11fad9880bc007