Distinguish internal uses of the serializer from PHP uses
Summary:
Add a new type to VariableSerializer called "Internal", which designates that
the serialization is being done for internal purposes (for example, storing in
the repo). Right now the internal type does the same thing as the normal
"serialize" type, but this allows us to treat these cases differently in the
future.

Reviewed By: alexeyt

Differential Revision: D6219307

fbshipit-source-id: cd6e71878c06d1168a03e1151eefab9e32ce2beb