Allow relabeling to be disabled in hh_single_compile
Summary:
Introduce a new option `Hack.Compiler.Relabel=0` that disables the renumbering of labels in the output of `hh_single_compile`.
The labels shouldn't matter for the purposes of `semdiff`, so we can improve performance a bit by avoiding the relabeling.
But it's useful to keep support for relabling in order that the output is syntactically closer to HHVM's for visual diff.

The default is to relabel, as normal.

Reviewed By: hubyrod

Differential Revision: D6231614

fbshipit-source-id: 9da8018c810c9d94326cd116d26b9629765b1617