Optimize/cleanup/rename asyncRetCtrl stub
Summary:
Rename this stub to asyncFuncRet and corresponding IR opcode to AsyncFuncRet,
as this stub is almost completely handling return from async functions.

Restructure the code in a way that makes it easier to add support for returning
from async functions to async generators.

Store the new kind/state when storing the result, as these data are right next
to each other in the memory.

Do not check whether a parent of a finished async function is in blocked state.
It always is. The purpose of this check was to check if the parent is an AFWH,
not whether it is blocked, but that is already being checked via AsioBlockable
kind.

Do not load data from AsioBlockable twice (testbim + load into reg). Load it
into reg once and use testbi{} instead of testbim{}. Do not clear bits that we
already checked are cleared when using it the second time.

Do not store result and state into the returning AFWH if it is going to be
destroyed. Instead, do the refcount check earlier and just directly free the
memory.

Structure the code in a way that besides rvmsp()/rvmtl() only one register
(parentBl) is alive during vcall{}s.

Reviewed By: mxw

Differential Revision: D6244390

fbshipit-source-id: 89208ee15aaf0f7ead9c171212c9132108055abe