Clean-up: Moving all settings to new parser env
Summary: Too many different places to set things in the lowerer. This diff cleans up the different types of settings / environments and should make it a lot easier to add flags. Ocaml's sometimes-`?`-means-`option`-sometimes-it-does-not mess now neads to be dealt with only and precisely once (in `make_env`).

Reviewed By: gregory-nisbet

Differential Revision: D6222693

fbshipit-source-id: 0602471193655a6063600650917a40f61d1371e0