Optional sexp sort before diff in `hh_single_parse`
Summary:
`Parser_hack` sometimes produces an `Ast` that is out-of-definition-order. When using `--parser compare` this adds a lot of noise, because the `Ast`s of both the legacy and full fidelity parsers are mapped to s-expressions in the order they appear.

For type-checking, definition order is invariant. For code-gen, because definitions are observable, it is not. The FFP reserves definition order. `hh_single_parse` should have low noise in the diffs it outputs, so lets just sort the output of the legacy parser.

Reviewed By: ericlippert

Differential Revision: D5923775

fbshipit-source-id: a1624718a96da3233a5efe429573580bd6232a4b