Fix semdiff syntactic output
Summary:
Motivation:
The syntactic diff was inverting the differences when reporting edits. For example, a deletion in the first file was reported as an addition.

Cause:
When parsing the arguments, we add the files in reverse order to !files.

Fix:
Before adding the two files to the `files` field in the `options` record, we swap the ordering.

Reviewed By: nbenton

Differential Revision: D6258698

fbshipit-source-id: 0546983bc926e2de94dc5343fa47c995e40b22ac