fix allowing overriding final methods through traits
Summary:
This diff is somewhat of a doozy; the original repro can be found in t14494468. There are a few things at play here that make this diff sort of subtle.

The original source of the bug is that is the finality check in `decl` only folds over the methods defined *on this class*[1].

To fix this, we move the check from `decl` to `typing`, otherwise we'd need to fold in all the trait methods and then discard the result so we don't screw up the downstream data structures in `Decl_defs`. That's pretty wasteful because we have the "merged" methods in typing.

In the `typing` check, we have to avoid expensive operations here, so we tag all the class members with the original source so we don't need to do set membership lookups to figure out if it's a method or a property. Also, since we check the class against the full merged impl, which includes interfaces and traits[2], we need to take care to only trigger this warning on actual class inheritance -- a class is allowed to override a trait-final method if that class is the one that actually uses the trait[3].

Overall, this check is quite cheap and introduces no new Hack errors.

Reviewed By: dlreeves

Differential Revision: D6137179

fbshipit-source-id: 16b5891cb70a94a2835696ecdd05f780baddcfcd