ban duplicate trait usage within a single class hierarchy
Summary:
We have decided from a language standpoint that reusing a trait is dangerous and should be banned. There are a lot of almost-disasters right now in the codebase -- if a single function in these reused traits became final, it would fatal at runtime and Hack would have no idea.

There is a bit of subtlety with how we track dependencies -- classes used in XHP attribute types are considered as "parents" of the current class so we typecheck in the right order. However, using a class as an XHP attribute never pulls traits into *this* class hierarchy, so we  need to consider them separately. Heap size only slightly increases because these two sets of dependencies are almost completely disjoint across the entire codebase (which makes sense).

In order to avoid computing set intersections repeatedly (which are linear w.r.t both sets with OCaml sets), we rely on the fact that single inheritance with no cycles means that if `size(old_deps) + size(parent_deps) > size(union(old_deps, parent_deps))`, then we have a reused dep that *must* be a trait. This allows us to defer the more expensive computations for when we are actually reporting such errors and this is about as cheap as we could possibly make this operation.

Reviewed By: dlreeves

Differential Revision: D6148365

fbshipit-source-id: 21f283af704035508a0b2d97bdc01ec522a4784c