Support merge-only units in hhbbc
Summary:
We'd never bothered with this because most things are persistent, so
most units only get loaded once. One downside is that this is the only
way to define persistent constants. Even that isn't much of an issue,
because almost all constant references have been resolved to their
value. But it still means that we have a normal-rds slot assigned to
each constant - which burns 24 bytes per const per thread, rather than
16 bytes per constant for the whole process.

Reviewed By: swtaarrs

Differential Revision: D6244219

fbshipit-source-id: d5bba5db7c3a51a82d54f11e39ff5ee4319ecb56