Remove unneeded full-fidelity-parser unit test
Summary:
When we were first building the FFP, we had an intern create a test case that randomly generated Hack programs in order to stress test the parser.

This was a good idea in the initial stages and found a lot of bugs, but eventually it became more trouble than it was worth because we had to keep the generation grammar generating legal Hack programs. This became hard as the parser got more sophisticated, and so I turned off those tests.

There were still a few of the supporting files hanging around, taking up time during builds and tests, and requiring updates when the FFP libraries changed. I've deleted them in this diff.
 
 
 

Differential Revision: D6262322

fbshipit-source-id: 21ce58197558caa21d0c1ee97c93f185828e4d3e