Continue functorizing the parser
Summary:
I have started to functorize the Syntax Tree, which is the entry point to the parser. (The syntax tree constructor takes a source text and provides a syntax tree; a syntax tree can give you, among other things, the root of the parse tree.)

Right now the parameter to the functor is ignored; in the next few diffs I'll make it use its parameter, and reorganize the error production code. We can then start switching out users of the syntax tree to generate positioned trees rather than minimal trees. 
 
 

Differential Revision: D6249153

fbshipit-source-id: 832c78d8ee1c8371fd7841b2908527f763e01c1d