Contextual inference for property initializers
Summary:
Use the declared type of a property as the expected type for its initializer expression. For example:
```
class C {
  public static vec<mixed> $my_prop = vec[2, "a"];
}
```
Right now, this won't have any discernible effect: lambdas cannot appear in the initializer because
they're not literal expressions, and so contextual inference won't be useful.

But in future we will be using the expected type to avoid unnecessary subtype or union operations e.g.
in the above example, if there are a large number of elements in the vector then it's useful to avoid
constructing a union type by simply doing (trivial) subtype tests against `mixed`.

Reviewed By: jamesjwu

Differential Revision: D6271076

fbshipit-source-id: 7589c10febfd69999d9353c67653927a9196a66f