Improve hash-table key exists checks.
Summary: Define a new FindType::Exists for findImpl, and a findForExists implemented using it, where instead of returning the pos we just return a boolean indicating whether the key was found at all. Use it in the MixedArray and SetArray Exists{Int,Str} methods instead of using find and checking if the returned pos was valid. This saves a number of instructions and reduces register pressure.

Reviewed By: mofarrell

Differential Revision: D6284221

fbshipit-source-id: 7d4bd69d6ed39eac284c0dd80e060d74d52e804b