Make wordsame faster for strings of 1-8 bytes.
Summary:
Explicitly unroll the nBits increment from the first iteration to the loop initialization, and push both increments below the "are we at the end of the string" check. Disassembly shows that this defers the increment of p1 until after we've decided to keep looping and jumped to the continue section of the loop.
Also, improve comments and make sure the fallback implementation used when FOLLY_SANITIZE_ADDRESS is defined has the same signature by makeing only the function body conditional on the macro.

Reviewed By: binliu19

Differential Revision: D6284227

fbshipit-source-id: e151027586347227229451b395b789be9bd2b39a