Fix gcc7 warning
Summary:
Without this, gcc7 warns that the malloc is allocating too much
memory. I think its cloning the code so we have one path that goes
through the loop, and one that does not, and then noting that the one
that doesn't would try to allocate a huge slab of memory.

This warning showed up in a lot of places, and I was able to disable
it everywhere else; but *this* one kept breaking the build (due to
-Werror).

Reviewed By: ricklavoie, alexeyt

Differential Revision: D6291577

fbshipit-source-id: d1e282abcb5f9954d423c9d1e14cf687d8f963d3