Set cloexec on listening sockets
Summary:
Sometimes (rarely) hh_client wedges itself on a Unix.select
before starting a new server (and after a version mismatch).

This seems to fix that.

Nitty gritty details:
After a Monitor exits due to version mismatch, there is a short
period where a client connecting to the socket doesn't fail (with
ECONNREFUSED) even though it should. One would expect that since the
connection is successful, it would eventually see either an EOF or
ECONNRESET when trying to read from the fd to indicate that the
listener on the other end has gone away. This doesn't happen either.

Instead what happens is a forever blocking "select".

My best guess is a race in the kernel
when cleaning up each process's open FDs, resulting in losing the expected
EOF (or equivalently ECONNRESET) signal to a pending connections.

To fix that, stop the problem earlier - we get the ECONNREFUSED that we
expect after the Monitor exits by preventing leakage of the listening
socket to child processes.

The Monitor spawns the Server process, which
spawns worker processes, each which have a copy of this
listening socket. So even though the Monitor exits, there may be
a child (the Server or Worker) still holding a listening FD, so
further connections to the socket after the Monitor dies
don't fail (with ECONNREFUSED) as would be expected.

Reviewed By: dabek

Differential Revision: D6296363

fbshipit-source-id: 71ee474dc84b5dc1a6bd4143136f1f7920d3024e