Refactor FPushClsMethod ops, add FPushClsMethodS and FPushClsMethodSD
Summary:
Refactor the FPushClsMethod bytecode ops. Remove FPushClsMethodF and add
FPushClsMethodS and FPushClsMethodSD. FPushClsMethodS and FPushClsMethodSD both
have a sub-op immediate, which determines if the class is "static", "self", or
"parent". These will be used instead of LateBoundCls/Self/Parent; FPushClsMethod
pairs, for example. Since the sub-op can be used to determine if the call is a
forwarding one or not, there's no need for a distinct FPushClsMethodF op.

The reason for doing this is we want to be able to distinguish between
dynamic/non-dynamic calls at the bytecode level. self::foo() isn't considered a
dynamic call, for example, but it uses the same bytecode push instruction as
$x::foo(). Now they'll use different bytecodes and we can distinguish them.

Reviewed By: mxw

Differential Revision: D6247840

fbshipit-source-id: 0e82d7bc19b8c8b051b0d009b1bc9f26b8d60091