Add support for distinguishing "dynamic" calls
Summary:
Add support to the runtime for distinguishing dynamic calls from static
ones. Generally speaking, a dynamic call is one whose target is not immediately
obvious from the FPush* op. Examples of dynamic calls:

- $x() (where $x is a string or array, but not an object)
- $x::foo()
- Foo:$x()
- $obj->$x()
- new $x()
- Invocations done via the runtime, such as in array_map or other builtins
  that execute callbacks (which aren't objects).

Calls to an object's __invoke() method are never considered dynamic (to support
closures).

To support this functionality, add a new flag bit to the ActRec. It is set by
the FPush* op if the op corresponds to a dynamic call (now that the bytecodes
have been refactored, there is no ambiguity). Inside the callee's prologue, if
the callee cares about being called dynamically, it can check the bit in the
ActRec and take the appropriate action.

To support this in the runtime, plumb through the dynamic flag from
vm_decode_function to where its used in invokeFuncImpl to set up the ActRec in
all the various places.

Add a dynamic flag a CufIter so it knows if the original callable was dynamic or
not (IE, if it was a string or array). DecodeCufIter will set this flag, and
FPushCufIter will move it into the ActRec. This is one of the few places where
the value of the dynamic flag may not be known statically at the site of the
FPush, so add the infrastructure to propagate its value. This means new IR
instructions to load/store the dynamic flag value in the CufIter, and the
alias-analysis and memory-effects support to load-elim the flag's value.

FCallAwait was assuming that none of the existing flags in the pre-live ActRec
were valid, and overwriting it with the FCallAwait flag. This is no longer true
with the dynamic flag, so change the logic to instead OR-in the FCallAwait flag,
not overwrite the existing values. This involved adding support for a "orlim"
instruction.

Add support to HHBBC to not optimize programs in a way that could potentially
change dynamic calls to static ones, in cases where the callee might care. For
example, don't transform a FPushFunc to a FPushFuncD, if the callee might care
that it was called dynamically. Likewise, don't attempt to fold such functions,
as their prologues might execute side-effects.

Finally, to test all this, add a new runtime option called
NoticeOnAllDynamicCalls. When this option is set, a dynamic call to *any*
function will raise a notice. In practice, this causes quite a bit of noise. Its
only meant for testing.

Reviewed By: markw65

Differential Revision: D6248926

fbshipit-source-id: 5716a58d46039b60fb18d31cb005a0e17dd036d9