Fix parsing of block scoped usings
Summary: In order to distinguish between block scoped vs function scoped using we used to call `parse_expression` and then check if subsequent token is semicolon. In some cases it is not correct - consider this case: `using ($a = new B()) {$c(); $d()}`. Here `parse_expression` will consume `($a = new B()) {$c()` as incomplete subscript expression, then check next token `;` and decide that it is function scoped using which is wrong.

Reviewed By: oulgen

Differential Revision: D6306819

fbshipit-source-id: 174d9695502f188def4374fada3f4f4421189f88