Continue functorizing the parser.
Summary:
The error checking pass and other code requires that "missing" nodes have positions in a positioned tree; when the tree was computed from a minimal tree, that was no problem. But now that the tree could be built directly from the parser, we have a problem. We build the tree from the bottom up, and compute the values that tag each syntax node *from the children*.  That value, in a positioned tree, includes the offset. But a "missing" node has no children, so we're going to need to pass the location in.

As several people have noted, we need a lightweight abstraction for positions that is better than tramping around a source text and offset all over the place. I'll try to implement that in a later diff. For now, I've plumbed in passing around a source text and offset so that when we construct a Missing node, we know where it is going. 
 

Differential Revision: D6278903

fbshipit-source-id: 4ddf90f36b948f9d878d4dd9aacda4396c9bebeb