New attribute AcceptDisposable
Summary:
In order to support benign built-in functions such as `var_dump` on disposables, we introduce
a new attribute `<<__AcceptDisposable>>` on function parameters, whose effect is to disable
the "escape" check on variables declared in a using clause.

The body of such a function is checked as though the parameters marked `<<__AcceptDisposable>>` are variables from a `using` clause i.e. they are subject to the same escape analysis.

Error messages for illegal escapes of Disposable are also improved.

Reviewed By: KendallHopkins

Differential Revision: D6297402

fbshipit-source-id: 358d031bb7c50cc7afdb712ade2363d92db22daf