Reject hints that implement IDisposable or IAsyncDisposable
Summary:
We want to make sure that disposable objects are correctly scoped using the new `using` statement.
Therefore we now outlaw the direct use of parameter type hints that are subtypes of `IDisposable` or `IAsyncDisposable`
unless explicitly flagged with the `<<__AcceptDisposable>>` attribute. This doesn't catch everything (e.g. generics)
but it should prevent most mistakes with disposables.

Reviewed By: KendallHopkins

Differential Revision: D6313311

fbshipit-source-id: 7263263e5598f1c38e3ba63950ff5ed90143432c