Be more succinct when stringifying JIT types
Summary:
When converting a JIT type to a string, we test if the type is exactly one of
the pre-defined types (union or not). If it is, we print it out. If it isn't, we
print it as a union of base non-union types. This can lead to large type
strings. For example, TArrLike|TObj is printed out as "
StaticArr|UncountedArr|CountedArr|StaticVec|UncountedVec|CountedVec|StaticDict|
UncountedDict|CountedDict|StaticKeyset|UncountedKeyset|CountedKeyset|TObj",
which is very unwieldy. If, instead, we decomposed the type into a minimal set
of union types, we'd be able to print this out as "ArrLike|Obj".

To accomplish this, build (statically) a list of all types (union or not),
sorted in decreasing size of their bit representation popcount. So, larger
unions come before smaller ones. Then, when building the string, use that list
to decompose the type. Since the larger unions come first, we should generate a
minimally sized string.

Reviewed By: markw65

Differential Revision: D6321825

fbshipit-source-id: be72b0eb6a175d16b1103ffdaeacb08cced17f7c