ast changes to support spread operator
Summary: The current AST representation for XHP attributes is `(id, expr)` pairs. This won't suffice for spreads, which are going to require special machinery to codegen and typecheck. In order to minimize the hackiness here, let's actually make the AST aware of the difference so every callsite handles it properly.

Differential Revision: D6316120

fbshipit-source-id: 737550f1ca5a51b94504a2416c2c7bfcf5102651