ffp changes for xhp spread
Summary: Implements the spread operator in the FFP. This involved a rename of some existing internals and new aggregate type to capture the different types of attributes that exist now. There are some XHP attribute hacks that we want to preserve for certain attribute name edge cases, but the spread operator doesn't have to deal with that and as a result, is pretty simple to parse.

Differential Revision: D6316613

fbshipit-source-id: 206b83b25c0791fdc5ab9633b807aabbf73e76be