HackC support for xhp spread operator
Summary:
This adds HackC support for the spread operator as described in the spec. XHP construction is emitted as a constructor with the attribute list as a `darray` with entries of the form `"attr-name" => expr`. Most of the XHP spread support happens at runtime, much like XHP attribute setting is deferred to the runtime implementation; the runtime will recognize the magic "...$" keys and take the array value to actually copy the attributes over. All that HackC has to do is keep track of how many spreads we have seen so far to generate unique array key names.

This approach accomplishes a few things: 1) it allows us to easily have the desired semantics (that is, there can be multiple spreads and their order and position matter for computing the final attribute list) and 2) this approach also fits nicely with how HHVM handles XHP internally (it's literally *parsed* as a constructor).

Reviewed By: oulgen

Differential Revision: D6317512

fbshipit-source-id: d64a7972f38811a0984503c92a251a398381fb49