Better inference for InstanceofD; Jmp[N]Z
Summary:
I noticed that we're getting a lot of

  Dup
  IsTypeC Null
  JmpZ
  Dup
  IsTypeC Null
  JmpZ

sequences in expanded genvas. It turns out that the first one was an
InstanceOfD check which got turned into an IsTypeC during the
optimization pass - by which time, its too late to kill the second
IsTypeC.

This adds the same is_opt checking we already do in InstanceOfD to the
(InstanceOfD,Jmp[N]Z) group.

Reviewed By: jano

Differential Revision: D6327983

fbshipit-source-id: 78f94104ccd3c559f75d89cdd7a48426f6ae3514