improved exceptional control flow tracking
Summary:
Change semdiff's model of exceptional control flow to be more
correct, and increase the disjunction limit to deal with complicated
examples. There's also a new toyexn.ml file that implements a model of
hhvm exceptions based on that in the documentation. The revised
version deals with one-sided control flow more liberally, but is also
a bit more restrictive in requiring catches to match up on the two
sides (because they affect the evaluation stack, popping from the
exception stack).

Reviewed By: vladima

Differential Revision: D6313246

fbshipit-source-id: 2fa143ec8c3e7f4733ff09c9581d717ed4929292