stop adding "XHP" to class extends list
Summary:
Today, the only way to tell if a given class is XHP is either to check if `tc_extends` contains `"XHP"` or check if the first character is a `:`. This is kind of weird, considering we do have the information on the parsed class data, and the `XHP` thing is a bug waiting to happen.

Luckily, the real XHP class is abstract and has no subclasses, or else Hack's inference of those constructors would be wrong[1].

At runtime, additionally, `<foo /> instanceof XHP` is not true. I couldn't really track down why this line was here, because Phabricator won't let me track the history properly from that far back.

I want to add typing utilities to verify if some type `T` is an XHP class, and this will make it less clowny to do so.

Reviewed By: dlreeves

Differential Revision: D6338858

fbshipit-source-id: 5ec5078f0b7488a6014da00b023acdf21336597f