Use single jsonrpc module instance
Summary:
I find ocaml functors hard\!

Prior to this diff I had coded them wrongly. I had coded it so clientLsp constructs one instance of Jsonrpc, and then Lsp_helpers constructs its own internal instance.

This was a problem, because the Jsonrpc module maintains state -- in particular, it maintains a list of all outstanding jsonrpc requests that hh_client has made to Nuclide and are awaiting responses. So when clientLsp asked Lsp_helpers to send a request, then it was stored in a Lsp_helpers-specific list of outstanding jsonrpc requests. And then when clientLsp received a response and tried to dispatch it, then it attempted to correlate the response with its own DIFFERENT list of outstanding jsonrpc requests.

The net effect was that when a user clicked "Restart" to restart the hh_server, we were unable to handle that button, and only emitted an telemetry/event notification saying "response to non-existent ID".

This diff fixes it up, by making Lsp_helpers module take Jsonrpc as a parameter. In this way both clientLsp and Lsp_helpers can share the same single instance of the Jsonrpc module.

Reviewed By: arxanas

Differential Revision: D6324460

fbshipit-source-id: 0b19d08c3a2e11d57fd9aa067ca0385b7b4b2a7d