Update typechecker to support type-hinted variadics within function type hints.
Summary:
This diff adds support for type-hinted variadics within function type hints.

Main changes:
In `decl_hint`, if the function hint contains a variadic parameter, we set the arity of `Tfun` to `Fvariadic` with either the actual type, or the any type `Hany`.

In `typing_subtype`:
If the function hint contains a variadic type hint and the function we pass in has a variadic parameter in the same position, we check that the type hint of the variadic in the function hint is a subtype of the variadic parameter on the function.
e.g
```
class A {}
class B extends A {}
function variadic_A(A ...$y): void {}
function hint_variadic_B((function(B...): void) $f): void {}
hint_variadic_B(fun('variadic_A')); //this should typecheck
```
If the function hint contains non-variadic type hints and the function we pass in has a variadic parameter starting at the same position, we check that the non-variadic type hints are subtypes of the type hint of the variadic parameter on the function.
e.g.
```
class A {}
class B extends A {}
function variadic_A(A ...$y): void {}
function hint_multiple_A_B((function(A, A, B, B, B): void) $f): void {}
hint_multiple_A_B(fun('variadic_A')); //this should typecheck
```

Reviewed By: andrewjkennedy

Differential Revision: D6086383

fbshipit-source-id: 01c772da651eab302e9854929674e8c09ddde8b2