Add Eval.AllowObjectDestructors and an attribute to bypass it
Summary:
The new option defaults to true except in one-bit refcount builds,
where it defaults to false and will be ignored if set to true. When set to
false, the runtime will fatal if you try to instantiate a class with a
__destruct() method. Users can set this to false to see if they're ready for
the future.

The <<__OptionalDestruct>> attribute can be added to __destruct() methods to
indicate to the runtime that they're safe to ignore. This allows instantiating
the class in question when Eval.AllowObjectDestructors = false.

Reviewed By: edwinsmith

Differential Revision: D6221409

fbshipit-source-id: 68545c72bf88f6e9970598c543a52ae8fe268918