inout parameters: function type hints
Summary:
Detailed specification: https://fb.quip.com/3utHAy5kEs8z

Add support for inout annotations in first-class function (closure) types. We require 'inout' to appear at both the definition and call sites, and the type checker needs to track the inout-ness of a parameter to enforce that it is used in a sound way.

Reviewed By: andrewjkennedy

Differential Revision: D6219212

fbshipit-source-id: 0bf3b301fa8c7df745a471a77f4aef35702eae07