Add env annotation to annotated AST
Summary:
Currently, the TAST defines a type `TAST.program`, which would seem to be the type of a TAST representing an entire file. This isn't actually the case, though--in order to make use of the TAST, it is necessary to use the "tast_tenv" type returned by `Typing.nast_to_tast_tenv` (i.e., a `(Tast.def * Typing_env.env) list`), since type variable substitutions derived during inference are only contained in the typing environment. Additionally, further substitutions are derived during the checking of method definitions, but are not represented in the returned environment at all.

To fix this, this stack adds an "Env" annotation to the annotated AST functor, so that the Typed AST can store typing environment information for subtrees which require it (currently, definitions of classes, methods, functions, and consts).

This diff adds EnvAnnotation to the AnnotatedAST functor and uses a unit annotation in the NAST and TAST to set up the necessary plumbing.

Reviewed By: andrewjkennedy

Differential Revision: D6255607

fbshipit-source-id: 1249b20ff29ce20b219f61b05a883eb93fd884ff